Move the wheel script lol
