--- Models/wheel.obj
@@ -1,152 +1,0 @@
-# Wheel
-v 0.5000000000000000 0.0000000000000000 0.0000000000000000
-v -0.5000000000000000 0.0000000000000000 0.0000000000000000
-v 0.5000000000000000 0.0000000000000000 1.0000000000000000
-v -0.5000000000000000 0.0000000000000000 1.0000000000000000
-v 0.5000000000000000 0.3090169943749474 0.9510565162951535
-v -0.5000000000000000 0.3090169943749474 0.9510565162951535
-v 0.5000000000000000 0.5877852522924731 0.8090169943749475
-v -0.5000000000000000 0.5877852522924731 0.8090169943749475
-v 0.5000000000000000 0.8090169943749475 0.5877852522924731
-v -0.5000000000000000 0.8090169943749475 0.5877852522924731
-v 0.5000000000000000 0.9510565162951535 0.3090169943749475
-v -0.5000000000000000 0.9510565162951535 0.3090169943749475
-v 0.5000000000000000 1.0000000000000000 0.0000000000000001
-v -0.5000000000000000 1.0000000000000000 0.0000000000000001
-v 0.5000000000000000 0.9510565162951536 -0.3090169943749473
-v -0.5000000000000000 0.9510565162951536 -0.3090169943749473
-v 0.5000000000000000 0.8090169943749475 -0.5877852522924730
-v -0.5000000000000000 0.8090169943749475 -0.5877852522924730
-v 0.5000000000000000 0.5877852522924732 -0.8090169943749473
-v -0.5000000000000000 0.5877852522924732 -0.8090169943749473
-v 0.5000000000000000 0.3090169943749475 -0.9510565162951535
-v -0.5000000000000000 0.3090169943749475 -0.9510565162951535
-v 0.5000000000000000 0.0000000000000001 -1.0000000000000000
-v -0.5000000000000000 0.0000000000000001 -1.0000000000000000
-v 0.5000000000000000 -0.3090169943749469 -0.9510565162951538
-v -0.5000000000000000 -0.3090169943749469 -0.9510565162951538
-v 0.5000000000000000 -0.5877852522924730 -0.8090169943749476
-v -0.5000000000000000 -0.5877852522924730 -0.8090169943749476
-v 0.5000000000000000 -0.8090169943749473 -0.5877852522924732
-v -0.5000000000000000 -0.8090169943749473 -0.5877852522924732
-v 0.5000000000000000 -0.9510565162951535 -0.3090169943749476
-v -0.5000000000000000 -0.9510565162951535 -0.3090169943749476
-v 0.5000000000000000 -1.0000000000000000 -0.0000000000000002
-v -0.5000000000000000 -1.0000000000000000 -0.0000000000000002
-v 0.5000000000000000 -0.9510565162951536 0.3090169943749472
-v -0.5000000000000000 -0.9510565162951536 0.3090169943749472
-v 0.5000000000000000 -0.8090169943749476 0.5877852522924729
-v -0.5000000000000000 -0.8090169943749476 0.5877852522924729
-v 0.5000000000000000 -0.5877852522924734 0.8090169943749473
-v -0.5000000000000000 -0.5877852522924734 0.8090169943749473
-v 0.5000000000000000 -0.3090169943749476 0.9510565162951535
-v -0.5000000000000000 -0.3090169943749476 0.9510565162951535
-v 0.5000000000000000 -0.0000000000000002 1.0000000000000000
-v -0.5000000000000000 -0.0000000000000002 1.0000000000000000
-vn 1.0000000000000000 0.0000000000000000 0.0000000000000000
-vn -1.0000000000000000 0.0000000000000000 0.0000000000000000
-vn 0.0000000000000000 0.1564344650402310 -0.9876883405951377
-vn -0.0000000000000000 -0.1564344650402310 -0.9876883405951377
-vn -0.0000000000000000 -0.4539904997395466 -0.8910065241883680
-vn -0.0000000000000000 -0.7071067811865476 -0.7071067811865476
-vn -0.0000000000000000 -0.8910065241883680 -0.4539904997395466
-vn -0.0000000000000000 -0.9876883405951377 -0.1564344650402310
-vn 0.0000000000000000 -0.9876883405951377 0.1564344650402306
-vn 0.0000000000000000 -0.8910065241883678 0.4539904997395469
-vn 0.0000000000000000 -0.7071067811865477 0.7071067811865474
-vn 0.0000000000000000 -0.4539904997395469 0.8910065241883678
-vn 0.0000000000000000 -0.1564344650402310 0.9876883405951377
-vn 0.0000000000000000 0.1564344650402305 0.9876883405951378
-vn 0.0000000000000000 0.4539904997395464 0.8910065241883681
-vn 0.0000000000000000 0.7071067811865476 0.7071067811865476
-vn 0.0000000000000000 0.8910065241883678 0.4539904997395469
-vn 0.0000000000000000 0.9876883405951377 0.1564344650402310
-vn 0.0000000000000000 0.9876883405951377 -0.1564344650402306
-vn 0.0000000000000000 0.8910065241883680 -0.4539904997395466
-vn 0.0000000000000000 0.7071067811865479 -0.7071067811865472
-vn 0.0000000000000000 0.4539904997395469 -0.8910065241883678
-vn 0.0000000000000000 0.1564344650402310 -0.9876883405951377
-f 1//1 3//1 39//1
-f 2//2 40//2 4//2
-f 39//3 3//3 4//3
-f 4//3 40//3 39//3
-f 1//1 5//1 3//1
-f 2//2 4//2 6//2
-f 3//4 5//4 6//4
-f 6//4 4//4 3//4
-f 1//1 7//1 5//1
-f 2//2 6//2 8//2
-f 5//5 7//5 8//5
-f 8//5 6//5 5//5
-f 1//1 9//1 7//1
-f 2//2 8//2 10//2
-f 7//6 9//6 10//6
-f 10//6 8//6 7//6
-f 1//1 11//1 9//1
-f 2//2 10//2 12//2
-f 9//7 11//7 12//7
-f 12//7 10//7 9//7
-f 1//1 13//1 11//1
-f 2//2 12//2 14//2
-f 11//8 13//8 14//8
-f 14//8 12//8 11//8
-f 1//1 15//1 13//1
-f 2//2 14//2 16//2
-f 13//9 15//9 16//9
-f 16//9 14//9 13//9
-f 1//1 17//1 15//1
-f 2//2 16//2 18//2
-f 15//10 17//10 18//10
-f 18//10 16//10 15//10
-f 1//1 19//1 17//1
-f 2//2 18//2 20//2
-f 17//11 19//11 20//11
-f 20//11 18//11 17//11
-f 1//1 21//1 19//1
-f 2//2 20//2 22//2
-f 19//12 21//12 22//12
-f 22//12 20//12 19//12
-f 1//1 23//1 21//1
-f 2//2 22//2 24//2
-f 21//13 23//13 24//13
-f 24//13 22//13 21//13
-f 1//1 25//1 23//1
-f 2//2 24//2 26//2
-f 23//14 25//14 26//14
-f 26//14 24//14 23//14
-f 1//1 27//1 25//1
-f 2//2 26//2 28//2
-f 25//15 27//15 28//15
-f 28//15 26//15 25//15
-f 1//1 29//1 27//1
-f 2//2 28//2 30//2
-f 27//16 29//16 30//16
-f 30//16 28//16 27//16
-f 1//1 31//1 29//1
-f 2//2 30//2 32//2
-f 29//17 31//17 32//17
-f 32//17 30//17 29//17
-f 1//1 33//1 31//1
-f 2//2 32//2 34//2
-f 31//18 33//18 34//18
-f 34//18 32//18 31//18
-f 1//1 35//1 33//1
-f 2//2 34//2 36//2
-f 33//19 35//19 36//19
-f 36//19 34//19 33//19
-f 1//1 37//1 35//1
-f 2//2 36//2 38//2
-f 35//20 37//20 38//20
-f 38//20 36//20 35//20
-f 1//1 39//1 37//1
-f 2//2 38//2 40//2
-f 37//21 39//21 40//21
-f 40//21 38//21 37//21
-f 1//1 41//1 39//1
-f 2//2 40//2 42//2
-f 39//22 41//22 42//22
-f 42//22 40//22 39//22
-f 1//1 43//1 41//1
-f 2//2 42//2 44//2
-f 41//23 43//23 44//23
-f 44//23 42//23 41//23
