+++ MonsterDuck/Assets/Models/wheel.py.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8cb440f22c06ca7449d30c770a93198f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
